backend/requirements.txt
asgiref==3.6.0
atomicwrites==1.4.1
attrs==23.1.0
black==23.3.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
cryptography==40.0.2
defusedxml==0.7.1
Django==4.1.3
django-colorfield==0.7.2
django-debug-toolbar==4.1.0
django-filter==23.2
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
djoser==2.2.0
exceptiongroup==1.1.1
flake8==6.0.0
gunicorn==20.1.0
idna==3.4
iniconfig==2.0.0
mccabe==0.7.0
mypy-extensions==1.0.0
oauthlib==3.2.2
packaging==23.1
pathspec==0.11.1
Pillow==9.5.0
platformdirs==3.5.1
pluggy==1.0.0
psycopg2==2.9.6
psycopg2-binary==2.9.6
py==1.11.0
pycodestyle==2.10.0
pycparser==2.21
pyflakes==3.0.1
PyJWT==2.7.0
pytest==6.2.5
pytest-django==4.5.2
pytest-pythonpath==0.7.4
python-dotenv==1.0.0
python-slugify==8.0.1
python3-openid==3.2.0
pytz==2023.3
requests==2.30.0
requests-oauthlib==1.3.1
social-auth-app-django==5.2.0
social-auth-core==4.4.2
sqlparse==0.4.4
text-unidecode==1.3
toml==0.10.2
tomli==2.0.1
tzdata==2023.3
urllib3==2.0.2
